feat: add new modal component
The creation of modals is necessary for the creation and editing of notes, which is why this element was developed.

In addition, some libraries have been added to the project to streamline the application development, such as:

Radix-ui -> Facilitating the development of modals.
Sonner -> Adding toasts for user notifications.
Date-fns -> Time formatting.
Lucide-react -> Icons.